Fix step tip styles overlapping during first animation (#28)
* Fix step tip styles overlapping during first animation

This PR fixes an issue that emerged after updating to ReacNative v0.64.0. It looks like native animations have improved, and give some of the styles are dynamic to create nice transitions/animations, the styles that come from from the implemented step tip can potentially overlap with the async animation styles. The fix is simply to check if the the animation styles are already present before rendering the tour tip.

* Fix layout callback issue
Fix tests